+++ packages/activemodel-adapter/lib/system/active-model-serializer.js
@@ -1,5 +1,6 @@
 import { singularize } from "ember-inflector";
 import RESTSerializer from "ember-data/serializers/rest-serializer";
+import normalizeModelName from "ember-data/system/normalize-model-name";
 /**
   @module ember-data
 */
@@ -137,17 +138,14 @@ var ActiveModelSerializer = RESTSerializer.extend({
   serializeHasMany: Ember.K,
 
   /**
-    Underscores the JSON root keys when serializing.
+   Underscores the JSON root keys when serializing.
 
-    @method serializeIntoHash
-    @param {Object} hash
-    @param {subclass of DS.Model} typeClass
-    @param {DS.Snapshot} snapshot
-    @param {Object} options
+    @method payloadKeyFromModelName
+    @param {String} modelName
+    @returns {String}
   */
-  serializeIntoHash: function(data, typeClass, snapshot, options) {
-    var root = underscore(decamelize(typeClass.modelName));
-    data[root] = this.serialize(snapshot, options);
+  payloadKeyFromModelName: function(modelName) {
+    return underscore(decamelize(modelName));
   },
 
   /**
@@ -268,11 +266,11 @@ var ActiveModelSerializer = RESTSerializer.extend({
           payloadKey = this.keyForAttribute(key, "deserialize");
           payload = hash[payloadKey];
           if (payload && payload.type) {
-            payload.type = this.typeForRoot(payload.type);
+            payload.type = this.modelNameFromPayloadKey(payload.type);
           } else if (payload && relationship.kind === "hasMany") {
             var self = this;
             forEach(payload, function(single) {
-              single.type = self.typeForRoot(single.type);
+              single.type = self.modelNameFromPayloadKey(single.type);
             });
           }
         } else {
@@ -289,11 +287,11 @@ var ActiveModelSerializer = RESTSerializer.extend({
       }, this);
     }
   },
-  typeForRoot: function(key) {
-    var normalized = camelize(singularize(key)).replace(/(^|\:)([A-Z])/g, function(match, separator, chr) {
+  modelNameFromPayloadKey: function(key) {
+    var convertedFromRubyModule = camelize(singularize(key)).replace(/(^|\:)([A-Z])/g, function(match, separator, chr) {
       return match.toLowerCase();
     }).replace('::', '/');
-    return this._super(normalized);
+    return normalizeModelName(convertedFromRubyModule);
   }
 });