+++ packages/ember-data/tests/unit/store/model-for-test.js
@@ -10,7 +10,7 @@ module("unit/store/model_for - DS.Store#modelFor", {
   setup: function() {
     env = setupStore({
       blogPost: DS.Model.extend(),
-      "blog-post": DS.Model.extend()
+      "blog.post": DS.Model.extend()
     });
     store = env.store;
     container = store.container;
@@ -31,7 +31,7 @@ test("when fetching factory from string, sets a normalized key as modelName", fu
   });
 
   equal(registry.normalize('some.post'), 'some-post', 'precond - container camelizes');
-  equal(store.modelFor('blog.post').modelName, "blog-post", "modelName is normalized to camelCase");
+  equal(store.modelFor("blog.post").modelName, "blog.post", "modelName is normalized to dasherized");
 });
 
 test("when fetching factory from string and dashing normalizer, sets a normalized key as modelName", function() {
@@ -39,7 +39,7 @@ test("when fetching factory from string and dashing normalizer, sets a normalize
     return dasherize(camelize(key));
   });
   equal(registry.normalize('some.post'), 'some-post', 'precond - container dasherizes');
-  equal(store.modelFor('blog.post').modelName, "blog-post", "modelName is normalized to camelCase");
+  equal(store.modelFor("blog.post").modelName, "blog.post", "modelName is normalized to dasherized");
 });
 
 test("when fetching something that doesn't exist, throws error", function() {